+++ lib/src/local_working_copy.rs
@@ -63,6 +63,8 @@ use crate::conflicts;
 use crate::conflicts::materialize_merge_result;
 use crate::conflicts::materialize_tree_value;
 use crate::conflicts::MaterializedTreeValue;
+#[cfg(unix)]
+use crate::file_util::check_executable_bit_support;
 use crate::file_util::check_symlink_support;
 use crate::file_util::try_symlink;
 #[cfg(feature = "watchman")]
@@ -90,7 +92,9 @@ use crate::op_store::WorkspaceId;
 use crate::repo_path::RepoPath;
 use crate::repo_path::RepoPathBuf;
 use crate::repo_path::RepoPathComponent;
+use crate::settings::ignore_executable_bit;
 use crate::settings::HumanByteSize;
+use crate::settings::UserSettings;
 use crate::store::Store;
 use crate::tree::Tree;
 use crate::working_copy::CheckoutError;
@@ -104,19 +108,115 @@ use crate::working_copy::WorkingCopy;
 use crate::working_copy::WorkingCopyFactory;
 use crate::working_copy::WorkingCopyStateError;
 
-#[cfg(unix)]
-type FileExecutableFlag = bool;
-#[cfg(windows)]
-type FileExecutableFlag = ();
+/// The executable bit for a filetype, potentially ignored.
+///
+/// On Windows there is no executable bit, so this will always be Ignore. On
+/// Unix it will usually be Exec(true|false), but may be ignored by a config
+/// value or by a check we run when we load the config if not specified.
+#[derive(Debug, Clone, Copy)]
+pub enum ExecFlag {
+    Exec(bool),
+    Ignore,
+}
+// Note: cannot derive Eq or PartialEq since a == b == c does not imply a == c.
+// e.g. `Exec(true) == Ignore == Exec(false)` but `Exec(true) != Exec(false)`
 
-#[derive(Debug, PartialEq, Eq, Clone)]
+impl ExecFlag {
+    fn matches(&self, other: &Self) -> bool {
+        match (self, other) {
+            (ExecFlag::Exec(a), ExecFlag::Exec(b)) => a == b,
+            // Always treat as equal if either is Ignore.
+            _ => true,
+        }
+    }
+
+    /// Create a bool in an environment where we can't check a IgnoreExec value.
+    fn from_bool_unchecked(executable: bool) -> Self {
+        if cfg!(unix) {
+            ExecFlag::Exec(executable)
+        } else {
+            ExecFlag::Ignore
+        }
+    }
+}
+
+/// Whether to ignore the executable bit when comparing files. The executable
+/// state is always ignored on Windows, but is respected by default on Unix and
+/// is ignored if we find that the filesystem doesn't support it or by user
+/// configuration.
+#[cfg(unix)]
+#[derive(Debug, Clone, Copy)]
+struct IgnoreExec(bool);
+#[cfg(windows)]
+#[derive(Debug, Clone, Copy)]
+struct IgnoreExec;
+
+impl IgnoreExec {
+    /// Load from user settings. If the setting is not given on Unix, then we
+    /// check whether executable bits are supported in the working copy's
+    /// filesystem and return true or false accordingly.
+    fn load_config(exec_config: Option<bool>, wc_path: &PathBuf) -> Self {
+        #[cfg(unix)] // check for executable support on Unix.
+        let ignore_exec =
+            IgnoreExec(exec_config.unwrap_or_else(|| !check_executable_bit_support(wc_path)));
+        #[cfg(windows)]
+        let (ignore_exec, _, _) = (IgnoreExec, exec_config, wc_path); // use the variables
+        ignore_exec
+    }
+
+    /// Push into an Option<bool> config value for roundtripping.
+    fn as_config(self) -> Option<bool> {
+        #[cfg(unix)]
+        let exec_config = Some(self.0);
+        #[cfg(windows)]
+        let (exec_config, _) = (None, self); // use the variable
+        exec_config
+    }
+
+    /// Resolve an executable bit into a flag for the FileType, potentially
+    /// ignoring it.
+    fn into_flag<F: Fn() -> bool>(self, is_executable: F) -> ExecFlag {
+        #[cfg(unix)]
+        let exec_flag = if self.0 {
+            ExecFlag::Ignore
+        } else {
+            ExecFlag::Exec(is_executable())
+        };
+        #[cfg(windows)]
+        let (exec_flag, _, _) = (ExecFlag::Ignore, self, is_executable); // use the variables
+        exec_flag
+    }
+
+    /// Convert a flag into the executable bit to write with a closure for a
+    /// default value.
+    fn exec_bit_to_write<F: Fn() -> bool>(self, exec_flag: ExecFlag, default: F) -> bool {
+        #[cfg(unix)]
+        let executable = match (self.0, exec_flag) {
+            (false, ExecFlag::Exec(executable)) => executable,
+            (true | false, _) => default(),
+        };
+        #[cfg(windows)]
+        let (executable, _, _) = (default(), self, exec_flag); // use the variables
+        executable
+    }
+}
+
+#[derive(Debug, Clone)]
 pub enum FileType {
-    Normal { executable: FileExecutableFlag },
+    Normal { exec_flag: ExecFlag },
     Symlink,
     GitSubmodule,
 }
 
-#[derive(Debug, PartialEq, Eq, Clone)]
+impl Default for FileType {
+    fn default() -> Self {
+        FileType::Normal {
+            exec_flag: ExecFlag::Exec(false),
+        }
+    }
+}
+
+#[derive(Debug, Clone)]
 pub struct FileState {
     pub file_type: FileType,
     pub mtime: MillisSinceEpoch,
@@ -130,30 +230,41 @@ impl FileState {
     /// Indicates that a file exists in the tree but that it needs to be
     /// re-stat'ed on the next snapshot.
     fn placeholder() -> Self {
-        #[cfg(unix)]
-        let executable = false;
-        #[cfg(windows)]
-        let executable = ();
         FileState {
-            file_type: FileType::Normal { executable },
+            file_type: FileType::default(),
             mtime: MillisSinceEpoch(0),
             size: 0,
         }
     }
 
-    fn for_file(executable: bool, size: u64, metadata: &Metadata) -> Self {
-        #[cfg(windows)]
-        let executable = {
-            // Windows doesn't support executable bit.
-            let _ = executable;
-        };
+    /// Filestate for a normal file..
+    fn for_file(size: u64, metadata: &Metadata, exec_flag: ExecFlag) -> Self {
         FileState {
-            file_type: FileType::Normal { executable },
+            file_type: FileType::Normal { exec_flag },
             mtime: mtime_from_metadata(metadata),
             size,
         }
     }
 
+    /// Whether this file state is compatible with another file state. The extra
+    /// complexity here comes from executable flags which always match on
+    /// Windows and might always match on Unix if ignore_exec is true.
+    fn matches(&self, other: &Self) -> bool {
+        use FileType::*;
+        let file_types_match = match (&self.file_type, &other.file_type) {
+            (GitSubmodule, GitSubmodule) | (Symlink, Symlink) => true,
+            (Normal { exec_flag: lhs }, Normal { exec_flag: rhs }) => lhs.matches(rhs),
+            _ => false,
+        };
+        file_types_match && self.mtime == other.mtime && self.size == other.size
+    }
+
+    /// Inverse of `self.matches(other)`.
+    fn differs(&self, other: &Self) -> bool {
+        !self.matches(other)
+    }
+
+    /// Filestate for a symlink.
     fn for_symlink(metadata: &Metadata) -> Self {
         // When using fscrypt, the reported size is not the content size. So if
         // we were to record the content size here (like we do for regular files), we
@@ -165,6 +276,7 @@ impl FileState {
         }
     }
 
+    /// Filestate for a git submodule.
     fn for_gitsubmodule() -> Self {
         FileState {
             file_type: FileType::GitSubmodule,
@@ -334,6 +446,7 @@ pub struct TreeState {
     sparse_patterns: Vec<RepoPathBuf>,
     own_mtime: MillisSinceEpoch,
     symlink_support: bool,
+    ignore_exec: IgnoreExec,
 
     /// The most recent clock value returned by Watchman. Will only be set if
     /// the repo is configured to use the Watchman filesystem monitor and
@@ -344,17 +457,14 @@ pub struct TreeState {
 fn file_state_from_proto(proto: &crate::protos::working_copy::FileState) -> FileState {
     let file_type = match proto.file_type() {
         crate::protos::working_copy::FileType::Normal => FileType::Normal {
-            executable: FileExecutableFlag::default(),
+            exec_flag: ExecFlag::from_bool_unchecked(false),
         },
-        #[cfg(unix)]
-        crate::protos::working_copy::FileType::Executable => FileType::Normal { executable: true },
-        // can exist in files written by older versions of jj
-        #[cfg(windows)]
-        crate::protos::working_copy::FileType::Executable => FileType::Normal { executable: () },
+        // can exist for Windows in files written by older versions of jj
+        crate::protos::working_copy::FileType::Executable => FileType::Normal {
+            exec_flag: ExecFlag::from_bool_unchecked(true),
+        },
         crate::protos::working_copy::FileType::Symlink => FileType::Symlink,
-        crate::protos::working_copy::FileType::Conflict => FileType::Normal {
-            executable: FileExecutableFlag::default(),
-        },
+        crate::protos::working_copy::FileType::Conflict => FileType::default(),
         crate::protos::working_copy::FileType::GitSubmodule => FileType::GitSubmodule,
     };
     FileState {
@@ -367,12 +477,10 @@ fn file_state_from_proto(proto: &crate::protos::working_copy::FileState) -> File
 fn file_state_to_proto(file_state: &FileState) -> crate::protos::working_copy::FileState {
     let mut proto = crate::protos::working_copy::FileState::default();
     let file_type = match &file_state.file_type {
-        #[cfg(unix)]
-        FileType::Normal { executable: false } => crate::protos::working_copy::FileType::Normal,
-        #[cfg(unix)]
-        FileType::Normal { executable: true } => crate::protos::working_copy::FileType::Executable,
-        #[cfg(windows)]
-        FileType::Normal { executable: () } => crate::protos::working_copy::FileType::Normal,
+        FileType::Normal {
+            exec_flag: ExecFlag::Exec(true),
+        } => crate::protos::working_copy::FileType::Executable,
+        FileType::Normal { exec_flag: _ } => crate::protos::working_copy::FileType::Normal,
         FileType::Symlink => crate::protos::working_copy::FileType::Symlink,
         FileType::GitSubmodule => crate::protos::working_copy::FileType::GitSubmodule,
     };
@@ -479,7 +587,7 @@ fn mtime_from_metadata(metadata: &Metadata) -> MillisSinceEpoch {
     )
 }
 
-fn file_state(metadata: &Metadata) -> Option<FileState> {
+fn file_state(metadata: &Metadata, ignore_exec: IgnoreExec) -> Option<FileState> {
     let metadata_file_type = metadata.file_type();
     let file_type = if metadata_file_type.is_dir() {
         None
@@ -487,13 +595,10 @@ fn file_state(metadata: &Metadata) -> Option<FileState> {
         Some(FileType::Symlink)
     } else if metadata_file_type.is_file() {
         #[cfg(unix)]
-        if metadata.permissions().mode() & 0o111 != 0 {
-            Some(FileType::Normal { executable: true })
-        } else {
-            Some(FileType::Normal { executable: false })
-        }
+        let exec_flag = ignore_exec.into_flag(|| metadata.permissions().mode() & 0o111 != 0);
         #[cfg(windows)]
-        Some(FileType::Normal { executable: () })
+        let exec_flag = ExecFlag::Ignore;
+        Some(FileType::Normal { exec_flag })
     } else {
         None
     };
@@ -567,18 +672,27 @@ impl TreeState {
         Box::new(PrefixMatcher::new(&self.sparse_patterns))
     }
 
+    /// Initialize an empty tree state and save it to the filesystem.
     pub fn init(
         store: Arc<Store>,
         working_copy_path: PathBuf,
         state_path: PathBuf,
+        exec_config: Option<bool>,
     ) -> Result<TreeState, TreeStateError> {
-        let mut wc = TreeState::empty(store, working_copy_path, state_path);
+        let mut wc = TreeState::empty(store, working_copy_path, state_path, exec_config);
         wc.save()?;
         Ok(wc)
     }
 
-    fn empty(store: Arc<Store>, working_copy_path: PathBuf, state_path: PathBuf) -> TreeState {
+    /// Create a new empty tree state for this working copy path.
+    fn empty(
+        store: Arc<Store>,
+        working_copy_path: PathBuf,
+        state_path: PathBuf,
+        exec_config: Option<bool>,
+    ) -> TreeState {
         let tree_id = store.empty_merged_tree_id();
+        let ignore_exec = IgnoreExec::load_config(exec_config, &working_copy_path);
         TreeState {
             store,
             working_copy_path,
@@ -589,18 +703,21 @@ impl TreeState {
             own_mtime: MillisSinceEpoch(0),
             symlink_support: check_symlink_support().unwrap_or(false),
             watchman_clock: None,
+            ignore_exec,
         }
     }
 
+    /// Load an existing tree state if present, or initialize an empty one.
     pub fn load(
         store: Arc<Store>,
         working_copy_path: PathBuf,
         state_path: PathBuf,
+        exec_config: Option<bool>,
     ) -> Result<TreeState, TreeStateError> {
         let tree_state_path = state_path.join("tree_state");
         let file = match File::open(&tree_state_path) {
             Err(ref err) if err.kind() == std::io::ErrorKind::NotFound => {
-                return TreeState::init(store, working_copy_path, state_path);
+                return TreeState::init(store, working_copy_path, state_path, exec_config);
             }
             Err(err) => {
                 return Err(TreeStateError::ReadTreeState {
@@ -611,7 +728,7 @@ impl TreeState {
             Ok(file) => file,
         };
 
-        let mut wc = TreeState::empty(store, working_copy_path, state_path);
+        let mut wc = TreeState::empty(store, working_copy_path, state_path, exec_config);
         wc.read(&tree_state_path, file)?;
         Ok(wc)
     }
@@ -624,6 +741,7 @@ impl TreeState {
         }
     }
 
+    /// Load the tree's data from the filesystem.
     fn read(&mut self, tree_state_path: &Path, mut file: File) -> Result<(), TreeStateError> {
         self.update_own_mtime();
         let mut buf = Vec::new();
@@ -655,6 +773,7 @@ impl TreeState {
         Ok(())
     }
 
+    /// Save the tree's data to the filesystem.
     #[allow(unknown_lints)] // XXX FIXME (aseipp): nightly bogons; re-test this occasionally
     #[allow(clippy::assigning_clones)]
     fn save(&mut self) -> Result<(), TreeStateError> {
@@ -855,7 +974,8 @@ impl TreeState {
                 file_states
                     .iter()
                     .filter(|(path, state)| {
-                        fsmonitor_matcher.matches(path) && state.file_type != FileType::GitSubmodule
+                        fsmonitor_matcher.matches(path)
+                            && !matches!(state.file_type, FileType::GitSubmodule)
                     })
                     .map(|(path, _state)| path.to_owned())
                     .collect()
@@ -959,7 +1079,7 @@ impl TreeState {
                 let path = dir.join(RepoPathComponent::new(name));
                 let maybe_current_file_state = file_states.get(&path);
                 if let Some(file_state) = &maybe_current_file_state {
-                    if file_state.file_type == FileType::GitSubmodule {
+                    if matches!(file_state.file_type, FileType::GitSubmodule) {
                         return Ok(());
                     }
                 }
@@ -994,7 +1114,7 @@ impl TreeState {
                                     });
                                 }
                             };
-                            if let Some(new_file_state) = file_state(&metadata) {
+                            if let Some(new_file_state) = file_state(&metadata, self.ignore_exec) {
                                 present_files_tx.send(tracked_path.to_owned()).ok();
                                 let update = self.get_updated_tree_value(
                                     tracked_path,
@@ -1008,7 +1128,7 @@ impl TreeState {
                                         .send((tracked_path.to_owned(), tree_value))
                                         .ok();
                                 }
-                                if new_file_state != current_file_state {
+                                if new_file_state.differs(&current_file_state) {
                                     file_states_tx
                                         .send((tracked_path.to_owned(), new_file_state))
                                         .ok();
@@ -1062,7 +1182,7 @@ impl TreeState {
                                 max_size: HumanByteSize(max_new_file_size),
                             });
                         }
-                        if let Some(new_file_state) = file_state(&metadata) {
+                        if let Some(new_file_state) = file_state(&metadata, self.ignore_exec) {
                             present_files_tx.send(path.clone()).ok();
                             let update = self.get_updated_tree_value(
                                 &path,
@@ -1074,7 +1194,10 @@ impl TreeState {
                             if let Some(tree_value) = update {
                                 tree_entries_tx.send((path.clone(), tree_value)).ok();
                             }
-                            if Some(&new_file_state) != maybe_current_file_state.as_ref() {
+                            if maybe_current_file_state
+                                .map(|fs| new_file_state.differs(&fs))
+                                .unwrap_or(true)
+                            {
                                 file_states_tx.send((path, new_file_state)).ok();
                             }
                         }
@@ -1147,7 +1270,8 @@ impl TreeState {
             Some(current_file_state) => {
                 // If the file's mtime was set at the same time as this state file's own mtime,
                 // then we don't know if the file was modified before or after this state file.
-                current_file_state == new_file_state && current_file_state.mtime < self.own_mtime
+                current_file_state.matches(new_file_state)
+                    && current_file_state.mtime < self.own_mtime
             }
         };
         if clean {
@@ -1166,8 +1290,8 @@ impl TreeState {
                 new_file_state.file_type.clone()
             };
             let new_tree_values = match new_file_type {
-                FileType::Normal { executable } => self
-                    .write_path_to_store(repo_path, &disk_path, &current_tree_values, executable)
+                FileType::Normal { exec_flag } => self
+                    .write_path_to_store(repo_path, &disk_path, &current_tree_values, exec_flag)
                     .block_on()?,
                 FileType::Symlink => {
                     let id = self
@@ -1190,24 +1314,19 @@ impl TreeState {
         repo_path: &RepoPath,
         disk_path: &Path,
         current_tree_values: &MergedTreeValue,
-        executable: FileExecutableFlag,
+        exec_flag: ExecFlag,
     ) -> Result<MergedTreeValue, SnapshotError> {
         // If the file contained a conflict before and is now a normal file on disk, we
         // try to parse any conflict markers in the file into a conflict.
         if let Some(current_tree_value) = current_tree_values.as_resolved() {
-            #[cfg(unix)]
-            let _ = current_tree_value; // use the variable
             let id = self.write_file_to_store(repo_path, disk_path).await?;
-            // On Windows, we preserve the executable bit from the current tree.
-            #[cfg(windows)]
-            let executable = {
-                let () = executable; // use the variable
-                if let Some(TreeValue::File { id: _, executable }) = current_tree_value {
-                    *executable
-                } else {
-                    false
-                }
-            };
+            // Use the given executable bit or return the current bit.
+            let executable =
+                self.ignore_exec
+                    .exec_bit_to_write(exec_flag, || match current_tree_value {
+                        Some(TreeValue::File { id: _, executable }) => *executable,
+                        _ => false,
+                    });
             Ok(Merge::normal(TreeValue::File { id, executable }))
         } else if let Some(old_file_ids) = current_tree_values.to_file_merge() {
             let content = fs::read(disk_path).map_err(|err| SnapshotError::Other {
@@ -1223,16 +1342,15 @@ impl TreeState {
             .block_on()?;
             match new_file_ids.into_resolved() {
                 Ok(file_id) => {
-                    // On Windows, we preserve the executable bit from the merged trees.
-                    #[cfg(windows)]
-                    let executable = {
-                        let () = executable; // use the variable
+                    // Use the given executable bit or preserve the executable
+                    // bit from the merged trees.
+                    let executable = self.ignore_exec.exec_bit_to_write(exec_flag, || {
                         if let Some(merge) = current_tree_values.to_executable_merge() {
-                            merge.resolve_trivial().copied().unwrap_or_default()
+                            merge.resolve_trivial().copied().unwrap_or(false)
                         } else {
                             false
                         }
-                    };
+                    });
                     Ok(Merge::normal(TreeValue::File {
                         id: file_id.unwrap(),
                         executable,
@@ -1269,7 +1387,7 @@ impl TreeState {
             message: format!("Failed to write file {}", disk_path.display()),
             err: err.into(),
         })?;
-        self.set_executable(disk_path, executable)?;
+        let exec_flag = self.set_executable_get_flag(disk_path, executable)?;
         // Read the file state from the file descriptor. That way, know that the file
         // exists and is of the expected type, and the stat information is most likely
         // accurate, except for other processes modifying the file concurrently (The
@@ -1277,7 +1395,7 @@ impl TreeState {
         let metadata = file
             .metadata()
             .map_err(|err| checkout_error_for_stat_error(err, disk_path))?;
-        Ok(FileState::for_file(executable, size, &metadata))
+        Ok(FileState::for_file(size, &metadata, exec_flag))
     }
 
     fn write_symlink(&self, disk_path: &Path, target: String) -> Result<FileState, CheckoutError> {
@@ -1316,22 +1434,28 @@ impl TreeState {
                 err: err.into(),
             })?;
         let size = conflict_data.len() as u64;
-        self.set_executable(disk_path, executable)?;
+        let exec_flag = self.set_executable_get_flag(disk_path, executable)?;
         let metadata = file
             .metadata()
             .map_err(|err| checkout_error_for_stat_error(err, disk_path))?;
-        Ok(FileState::for_file(executable, size, &metadata))
+        Ok(FileState::for_file(size, &metadata, exec_flag))
     }
 
-    #[cfg_attr(windows, allow(unused_variables))]
-    fn set_executable(&self, disk_path: &Path, executable: bool) -> Result<(), CheckoutError> {
+    /// Maybe set the executable bit and return the flag or an error. This is a
+    /// no-op on Windows.
+    fn set_executable_get_flag(
+        &self,
+        disk_path: &Path,
+        executable: bool,
+    ) -> Result<ExecFlag, CheckoutError> {
+        let exec_flag = self.ignore_exec.into_flag(|| executable);
         #[cfg(unix)]
-        {
+        if let ExecFlag::Exec(executable) = exec_flag {
             let mode = if executable { 0o755 } else { 0o644 };
             fs::set_permissions(disk_path, fs::Permissions::from_mode(mode))
                 .map_err(|err| checkout_error_for_stat_error(err, disk_path))?;
-        }
-        Ok(())
+        };
+        Ok(exec_flag)
     }
 
     pub fn check_out(&mut self, new_tree: &MergedTree) -> Result<CheckoutStats, CheckoutError> {
@@ -1513,10 +1637,9 @@ impl TreeState {
             } else {
                 let file_type = match after.into_resolved() {
                     Ok(value) => match value.unwrap() {
-                        #[cfg(unix)]
-                        TreeValue::File { id: _, executable } => FileType::Normal { executable },
-                        #[cfg(windows)]
-                        TreeValue::File { .. } => FileType::Normal { executable: () },
+                        TreeValue::File { id: _, executable } => FileType::Normal {
+                            exec_flag: self.ignore_exec.into_flag(|| executable),
+                        },
                         TreeValue::Symlink(_id) => FileType::Symlink,
                         TreeValue::Conflict(_id) => {
                             panic!("unexpected conflict entry in diff at {path:?}");
@@ -1529,11 +1652,17 @@ impl TreeState {
                             panic!("unexpected tree entry in diff at {path:?}");
                         }
                     },
-                    Err(_values) => {
-                        // TODO: Try to set the executable bit based on the conflict
-                        FileType::Normal {
-                            executable: FileExecutableFlag::default(),
+                    // TODO: Try to set the executable bit based on the conflict
+                    Err(values) => {
+                        let mut file_type = FileType::default();
+                        for value in values.adds().flatten() {
+                            // Use the *last* added filetype from the merge
+                            if let TreeValue::File { id: _, executable } = value {
+                                let exec_flag = self.ignore_exec.into_flag(|| *executable);
+                                file_type = FileType::Normal { exec_flag };
+                            }
                         }
+                        file_type
                     }
                 };
                 let file_state = FileState {
@@ -1577,6 +1706,7 @@ pub struct LocalWorkingCopy {
     state_path: PathBuf,
     checkout_state: OnceCell<CheckoutState>,
     tree_state: OnceCell<TreeState>,
+    ignore_exec: IgnoreExec,
 }
 
 impl WorkingCopy for LocalWorkingCopy {
@@ -1617,6 +1747,7 @@ impl WorkingCopy for LocalWorkingCopy {
             // TODO: It's expensive to reload the whole tree. We should copy it from `self` if it
             // hasn't changed.
             tree_state: OnceCell::new(),
+            ignore_exec: self.ignore_exec,
         };
         let old_operation_id = wc.operation_id().clone();
         let old_tree_id = wc.tree_id()?.clone();
@@ -1645,6 +1776,7 @@ impl LocalWorkingCopy {
         state_path: PathBuf,
         operation_id: OperationId,
         workspace_id: WorkspaceId,
+        settings: &UserSettings,
     ) -> Result<LocalWorkingCopy, WorkingCopyStateError> {
         let proto = crate::protos::working_copy::Checkout {
             operation_id: operation_id.to_bytes(),
@@ -1656,19 +1788,24 @@ impl LocalWorkingCopy {
             .open(state_path.join("checkout"))
             .unwrap();
         file.write_all(&proto.encode_to_vec()).unwrap();
-        let tree_state =
-            TreeState::init(store.clone(), working_copy_path.clone(), state_path.clone()).map_err(
-                |err| WorkingCopyStateError {
-                    message: "Failed to initialize working copy state".to_string(),
-                    err: err.into(),
-                },
-            )?;
+        let tree_state = TreeState::init(
+            store.clone(),
+            working_copy_path.clone(),
+            state_path.clone(),
+            ignore_executable_bit(settings.config()),
+        )
+        .map_err(|err| WorkingCopyStateError {
+            message: "Failed to initialize working copy state".to_string(),
+            err: err.into(),
+        })?;
+        let ignore_exec = tree_state.ignore_exec;
         Ok(LocalWorkingCopy {
             store,
             working_copy_path,
             state_path,
             checkout_state: OnceCell::new(),
             tree_state: OnceCell::with_value(tree_state),
+            ignore_exec,
         })
     }
 
@@ -1676,13 +1813,17 @@ impl LocalWorkingCopy {
         store: Arc<Store>,
         working_copy_path: PathBuf,
         state_path: PathBuf,
+        settings: &UserSettings,
     ) -> LocalWorkingCopy {
+        let exec_config = ignore_executable_bit(settings.config());
+        let ignore_exec = IgnoreExec::load_config(exec_config, &working_copy_path);
         LocalWorkingCopy {
             store,
             working_copy_path,
             state_path,
             checkout_state: OnceCell::new(),
             tree_state: OnceCell::new(),
+            ignore_exec,
         }
     }
 
@@ -1731,6 +1872,7 @@ impl LocalWorkingCopy {
                     self.store.clone(),
                     self.working_copy_path.clone(),
                     self.state_path.clone(),
+                    self.ignore_exec.as_config(),
                 )
             })
             .map_err(|err| WorkingCopyStateError {
@@ -1793,6 +1935,7 @@ impl WorkingCopyFactory for LocalWorkingCopyFactory {
         state_path: PathBuf,
         operation_id: OperationId,
         workspace_id: WorkspaceId,
+        settings: &UserSettings,
     ) -> Result<Box<dyn WorkingCopy>, WorkingCopyStateError> {
         Ok(Box::new(LocalWorkingCopy::init(
             store,
@@ -1800,6 +1943,7 @@ impl WorkingCopyFactory for LocalWorkingCopyFactory {
             state_path,
             operation_id,
             workspace_id,
+            settings,
         )?))
     }
 
@@ -1808,11 +1952,13 @@ impl WorkingCopyFactory for LocalWorkingCopyFactory {
         store: Arc<Store>,
         working_copy_path: PathBuf,
         state_path: PathBuf,
+        settings: &UserSettings,
     ) -> Result<Box<dyn WorkingCopy>, WorkingCopyStateError> {
         Ok(Box::new(LocalWorkingCopy::load(
             store,
             working_copy_path,
             state_path,
+            settings,
         )))
     }
 }
@@ -1979,12 +2125,18 @@ mod tests {
         RepoPath::from_internal_string(value)
     }
 
+    // Only for convenience in these tests. File states are *not* transitively
+    // equal (due to ExecFlag), so we should not implement PartialEq generally.
+    impl PartialEq for FileState {
+        fn eq(&self, other: &Self) -> bool {
+            self.matches(other)
+        }
+    }
+
     #[test]
     fn test_file_states_merge() {
         let new_state = |size| FileState {
-            file_type: FileType::Normal {
-                executable: FileExecutableFlag::default(),
-            },
+            file_type: FileType::default(),
             mtime: MillisSinceEpoch(0),
             size,
         };
@@ -2030,9 +2182,7 @@ mod tests {
     #[test]
     fn test_file_states_lookup() {
         let new_state = |size| FileState {
-            file_type: FileType::Normal {
-                executable: FileExecutableFlag::default(),
-            },
+            file_type: FileType::default(),
             mtime: MillisSinceEpoch(0),
             size,
         };